+++ backend/app/__init__.py
@@ -1,0 +1,9 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+from config import Config
+
+app = Flask(__name__)
+app.config.from_object(Config)
+db = SQLAlchemy(app)
+
+from app import routes, models, helpers
